+++ semcore/color-picker/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [2.25.0-prerelease.1] - 2024-02-13
+## [2.25.1] - 2024-02-16
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/dropdown` [4.19.2 ~> 4.19.3]).
+
+## [2.25.0] - 2024-02-13
 
 ### Changed